NFC, move variable placement logic to ExtractionUtils
Prep for rdar://32793311
