+++ lib/Tooling/Refactor/ExtractRepeatedExpressionIntoVariable.cpp
@@ -19,7 +19,6 @@
 #include "clang/AST/ASTContext.h"
 #include "clang/AST/Expr.h"
 #include "clang/AST/RecursiveASTVisitor.h"
-#include "llvm/Support/SaveAndRestore.h"
 
 using namespace clang;
 using namespace clang::tooling;
@@ -234,94 +233,6 @@ clang::tooling::initiateExtractRepeatedExpressionIntoVariableOperation(
   return Result;
 }
 
-namespace {
-
-/// Checks if a set of expressions is directly contained in some AST region.
-class StmtReachabilityChecker
-    : public RecursiveASTVisitor<StmtReachabilityChecker> {
-  const llvm::SmallPtrSetImpl<const Stmt *> &Expressions;
-  unsigned Count = 0;
-
-  StmtReachabilityChecker(
-      const llvm::SmallPtrSetImpl<const Stmt *> &Expressions)
-      : Expressions(Expressions) {}
-
-  bool areAllExpressionsReached() const { return Count == Expressions.size(); }
-
-public:
-  bool VisitStmt(const Stmt *S) {
-    if (Expressions.count(S)) {
-      ++Count;
-      if (areAllExpressionsReached())
-        return false;
-    }
-    return true;
-  }
-
-  static bool areAllExpressionsReachableFrom(
-      CompoundStmt *S, const llvm::SmallPtrSetImpl<const Stmt *> &Expressions) {
-    StmtReachabilityChecker Checker(Expressions);
-    Checker.TraverseStmt(S);
-    return Checker.areAllExpressionsReached();
-  }
-};
-
-/// Figures out where the extracted variable should go.
-class ExtractedVariableInsertionLocFinder
-    : public RecursiveASTVisitor<ExtractedVariableInsertionLocFinder> {
-  llvm::SmallPtrSet<const Stmt *, 4> Expressions;
-  llvm::SmallVector<std::pair<CompoundStmt *, const Stmt *>, 4>
-      InsertionCandidateStack;
-  bool IsPrevCompoundStmt = false;
-
-public:
-  SourceLocation Loc;
-
-  /// Initializes the insertion location finder using the set of duplicate
-  /// \p Expressions from one function.
-  ExtractedVariableInsertionLocFinder(ArrayRef<const Expr *> Expressions) {
-    for (const Expr *E : Expressions)
-      this->Expressions.insert(E);
-  }
-
-  bool TraverseStmt(Stmt *S) {
-    if (!S)
-      return RecursiveASTVisitor::TraverseStmt(S);
-    if (IsPrevCompoundStmt && !InsertionCandidateStack.empty())
-      InsertionCandidateStack.back().second = S;
-    llvm::SaveAndRestore<bool> IsPrevCompoundStmtTracker(IsPrevCompoundStmt,
-                                                         false);
-    if (auto *CS = dyn_cast<CompoundStmt>(S)) {
-      IsPrevCompoundStmt = true;
-      InsertionCandidateStack.emplace_back(CS, nullptr);
-      return RecursiveASTVisitor::TraverseStmt(S);
-    }
-    return RecursiveASTVisitor::TraverseStmt(S);
-  }
-
-  bool VisitStmt(const Stmt *S) {
-    if (Expressions.count(S)) {
-      // The insertion location should be in the first compound statement that
-      // includes all of the expressions as descendants as we want the new
-      // variable to be visible to all uses.
-      for (auto I = InsertionCandidateStack.rbegin(),
-                E = InsertionCandidateStack.rend();
-           I != E; ++I) {
-        if (StmtReachabilityChecker::areAllExpressionsReachableFrom(
-                I->first, Expressions) &&
-            I->second) {
-          Loc = I->second->getLocStart();
-          break;
-        }
-      }
-      return false;
-    }
-    return true;
-  }
-};
-
-} // end anonymous namespace
-
 static StringRef nameForExtractedVariable(const Expr *E) {
   auto SuggestedName = extract::nameForExtractedVariable(E);
   if (!SuggestedName)
@@ -336,9 +247,10 @@ ExtractRepeatedExpressionIntoVariableOperation::perform(
   std::vector<RefactoringReplacement> Replacements;
 
   const SourceManager &SM = Context.getSourceManager();
-  ExtractedVariableInsertionLocFinder LocFinder(DuplicateExpressions);
-  LocFinder.TraverseDecl(const_cast<Decl *>(ParentDecl));
-  if (LocFinder.Loc.isInvalid())
+  SourceLocation InsertionLoc =
+      extract::locationForExtractedVariableDeclaration(DuplicateExpressions,
+                                                       ParentDecl, SM);
+  if (InsertionLoc.isInvalid())
     return llvm::make_error<RefactoringOperationError>(
         "no appropriate insertion location found");
 
@@ -356,9 +268,6 @@ ExtractRepeatedExpressionIntoVariableOperation::perform(
   OS << " = ";
   E->printPretty(OS, /*Helper=*/nullptr, Context.getPrintingPolicy());
   OS << ";\n";
-  SourceLocation InsertionLoc = LocFinder.Loc;
-  if (InsertionLoc.isMacroID())
-    InsertionLoc = SM.getExpansionLoc(InsertionLoc);
   Replacements.emplace_back(SourceRange(InsertionLoc, InsertionLoc), OS.str());
 
   // Replace the duplicates with a reference to the variable.
